Agentic AI Q4/Self_Practise/s.k/Class three/pyproject.toml
[project]
name = "class-three"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.6.5",
    "dotenv>=0.9.9",
    "openai-agents>=0.2.4",
]
